requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
Django==3.2.7
django-request-logging==0.7.3
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==3.2
pycparser==2.20
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
